feat: implement alert publishing for archiving failures and add SNS configuration
